usb_poll: Match up signatures

+++ src/unfloader_debug/usb.c
@@ -620,7 +620,7 @@ char usb_write(int datatype, const void* data, int size)
     @return The data header, or 0
 ==============================*/
 
-u32 usb_poll(void)
+long unsigned int usb_poll(void)
 {
     // If no debug cart exists, stop
     if (usb_cart == CART_NONE)
